Smart Wallets: Isolate Web3Auth Specific Code Into Adapter Package (#570)
* isolate web3auth specific code into adapter package

* adjust tsconfig

* Rename + tidy
