+++ packages/client/wallets/smart-wallet/src/types/Config.ts
@@ -1,8 +1,7 @@
-import { TORUS_NETWORK_TYPE } from "@web3auth/single-factor-auth";
-import { EntryPoint } from "permissionless/types/entrypoint";
-import { EIP1193Provider, LocalAccount } from "viem";
+import type { EntryPoint } from "permissionless/types/entrypoint";
+import type { EIP1193Provider, LocalAccount } from "viem";
 
-import { SupportedEntryPointVersion } from "./internal";
+import type { SupportedEntryPointVersion } from "./internal";
 
 export type SmartWalletSDKInitParams = {
     clientApiKey: string;
@@ -14,14 +13,6 @@ export type UserParams = {
      * that is specified in the project settings (typically `sub`).
      */
     id: string;
-    jwt: string;
-};
-
-export type Web3AuthSigner = {
-    type: "WEB3_AUTH";
-    clientId: string;
-    verifierId: string;
-    web3AuthNetwork: TORUS_NETWORK_TYPE;
     jwt: string;
 };
 
@@ -39,7 +30,7 @@ export type PasskeySigner = {
     passkeyName: string;
 };
 
-export type EOASigner = EIP1193Provider | Web3AuthSigner | ViemAccount;
+export type EOASigner = EIP1193Provider | ViemAccount;
 export interface WalletConfig {
     signer: EOASigner | PasskeySigner;
 }
