+++ packages/client/wallets/web3auth-adapter/.gitignore
@@ -1,0 +1,5 @@
+dist
+node_modules/
+/test-results/
+/blob-report/
+
