+++ packages/client/wallets/web3auth-adapter/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@crossmint/client-sdk-smart-wallet",
+    "name": "@crossmint/client-sdk-smart-wallet-web3auth-adapter",
     "version": "0.1.0",
     "repository": "https://github.com/Crossmint/crossmint-sdk",
     "license": "Apache-2.0",
@@ -29,18 +29,12 @@
     },
     "dependencies": {
         "@crossmint/common-sdk-base": "0.0.13",
-        "@datadog/browser-logs": "4.42.2",
-        "@zerodev/ecdsa-validator": "5.3.1",
-        "@zerodev/passkey-validator": "5.3.5",
-        "@zerodev/sdk": "5.3.3",
-        "@zerodev/webauthn-key": "5.3.1",
-        "permissionless": "0.1.36",
-        "uuid": "10.0.0",
+        "@web3auth/base": "7.3.2",
+        "@web3auth/ethereum-provider": "7.3.2",
+        "@web3auth/single-factor-auth": "7.3.0",
         "viem": "2.17.5"
     },
     "devDependencies": {
-        "@playwright/test": "^1.41.2",
-        "@types/node": "20.12.7",
-        "@types/uuid": "10.0.0"
+        "@types/node": "20.12.7"
     }
 }
