+++ packages/client/wallets/web3auth-adapter/src/auth.test.ts
