+++ packages/client/wallets/web3auth-adapter/src/auth.ts
@@ -1,8 +1,3 @@
-import { logError } from "@/services/logging";
-
-import { SCW_SERVICE } from "./constants";
-import { errorToJSON } from "./log";
-
 export const parseToken = (token: any) => {
     try {
         const base64Url = token.split(".")[1];
@@ -11,10 +6,6 @@ export const parseToken = (token: any) => {
             typeof window !== "undefined" ? window.atob(base64) : Buffer.from(base64, "base64").toString();
         return JSON.parse(jsonPayload || "");
     } catch (err) {
-        logError("[PARSE_TOKEN] - ERROR", {
-            service: SCW_SERVICE,
-            error: errorToJSON(err),
-        });
         console.error("Error while parsing token");
         throw err;
     }
