+++ packages/client/wallets/web3auth-adapter/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../../../tsconfig.base.json",
+    "compilerOptions": {
+        "rootDir": "src",
+        "baseUrl": ".",
+        "paths": {
+            "@/*": ["src/*"]
+        }
+    },
+    "include": ["src/**/*"]
+}
